requirements.txt
fastapi
uvicorn
pydantic
python-jose[cryptography]
python-multipart
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-dotenv
email-validator
pytest
pytest-cov
httpx